Bump version to 3.7.1 (#63)
## Problem

I didn't realize I needed to manually bump the version for the changes in #62. The tasks we normally use to do this are from a private Stitch Fix gem (`y`), which is not included in `stitches`.

## Solution

Bump the version from `3.6.1` to `3.7.1`, since it's a minor release. After deploying, I will run the `rake release` task.