Working auth for freshly generated APIs
The default Rails 5 template uses the :exception strategy for forgery
protection; it raises an exception if a request doesn't contain a
valid CSRF token.

This is a good default for user-facing apps, but for services it means
that a freshly generated app will turn away all API requests---even
ones with valid API keys.

This commit configures stitches-generated APIs to use the
`:null_session` CSRF strategy for `ApiController` descendants---but
retains the default `:exception` strategy for all other controllers.