Remove Apitome as forced dependency
This removes Apitome (hosts docs generated by rspec_api_documentation)
from the generator and gem and thereby permits applications that use
stitches to opt-out of hosting the documentation.

Hosting the documentation in production requires an otherwise API-only
application to include dependencies such as ActionView and Nokogiri.
Viewing documentation is a relatively infrequent thing and only helps
developers who also have access to Github, so the suggested alternative
is to view the generated documentation locally:

    # Generate the docs if they don't already exist
    $ RAILS_ENV=test bin/rails docs:generate
    $ open docs/api/index.html