Run generators within application bundle
The test is creating a mock application, but isn't including the
"stitches" gem in the Gemfile in as the directions indicate, so even
though the generator is adding "apitome" to the application's Gemfile,
it's still necessary to have "apitome" as a runtime dependency of the
gem.

The gem inserts "apitome" and "rspec-rails" (now adds, rather than assuming it
already there) and within the gem context of the application executes
the generators and updates the files.