Cache api_keys
Currently have request to an internal service requires at least one
DB hit. For some services this is noise, for others it's the only
reason they have a databse. In a few cases, API key look up is
what the database is spending most of its time on.

Since this data is effectively immutable it's a great candidate for
caching. All apps would benefit from this.

This caches API keys in memory using a bounded LRU cache.